+++ CHANGELOG.md
@@ -56,6 +56,7 @@ S3 BACKEND:
 * Adds support for the `assume_role_with_web_identity` block. ([#689](https://github.com/opentofu/opentofu/issues/689))
 * Adds support for account whitelisting using the `forbidden_account_ids` and `allowed_account_ids` arguments. ([#699](https://github.com/opentofu/opentofu/issues/699))
 * Adds the `custom_ca_bundle` argument. ([#689](https://github.com/opentofu/opentofu/issues/689))
+* Adds support for the `sts_region` argument. ([#695](https://github.com/opentofu/opentofu/issues/695))
 
 ## Previous Releases